good night

+++ libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: moboigui <moboigui@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/13 20:20:22 by moboigui          #+#    #+#             */
-/*   Updated: 2022/11/13 20:20:23 by moboigui         ###   ########.fr       */
+/*   Updated: 2022/11/26 21:53:16 by moboigui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char const *s)
+#include <stddef.h>
+
+size_t	ft_strlen(char const *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
